(Logistic Regression Optimization)
Logistic regression is a supervised machine learning algorithm used for
classification tasks where the goal is to predict the probability that an instance
belongs to a given class or not. The probability of the class is first calculated by
multiplying the input features with its corresponding weights then the bias term is
added. After that, the output (Z) is mapped into a sigmoid function to output a value
between 0-1.

z = wx + b
g(z) = 1 / 1 - e^-z

use the 4 types optimizters :
-> Gradient Descent
-> Batch Gradient Descent
-> Stochastic Gradient Descent
-> Mini Batch Gradient Descent